Rename files for consistency (#1247)
