+++ client/public/images/status/bad-omen.png
