+++ client/public/images/status/royal-protection.png
