+++ common/cards/advent-of-tcg/attach/brewing-stand.ts
@@ -1,7 +1,7 @@
 import {CardComponent, ObserverComponent} from '../../../components'
 import query from '../../../components/query'
 import {GameModel} from '../../../models/game-model'
-import {flipCoin} from '../../../utils/coinFlips'
+import {flipCoin} from '../../../utils/coin-flips'
 import {attach} from '../../defaults'
 import {Attach} from '../../types'
