+++ common/cards/advent-of-tcg/hermits/monkeyfarm-rare.ts
@@ -6,7 +6,7 @@ import {
 import query from '../../../components/query'
 import {GameModel} from '../../../models/game-model'
 import {afterAttack} from '../../../types/priorities'
-import {flipCoin} from '../../../utils/coinFlips'
+import {flipCoin} from '../../../utils/coin-flips'
 import {hermit} from '../../defaults'
 import {Hermit} from '../../types'
