+++ common/cards/advent-of-tcg/hermits/pharaoh-rare.ts
@@ -7,7 +7,7 @@ import query from '../../../components/query'
 import {ReadonlyAttackModel} from '../../../models/attack-model'
 import {GameModel} from '../../../models/game-model'
 import {afterAttack, beforeAttack} from '../../../types/priorities'
-import {flipCoin} from '../../../utils/coinFlips'
+import {flipCoin} from '../../../utils/coin-flips'
 import {hermit} from '../../defaults'
 import {Hermit} from '../../types'
