+++ common/cards/advent-of-tcg/single-use/tnt-minecart.ts
@@ -2,7 +2,7 @@ import {CardComponent, ObserverComponent} from '../../../components'
 import {GameModel} from '../../../models/game-model'
 import {beforeAttack} from '../../../types/priorities'
 import {applySingleUse} from '../../../utils/board'
-import {flipCoin} from '../../../utils/coinFlips'
+import {flipCoin} from '../../../utils/coin-flips'
 import {singleUse} from '../../defaults'
 import {SingleUse} from '../../types'
