+++ common/cards/hermits/ethoslab-rare.ts
@@ -7,7 +7,7 @@ import query from '../../components/query'
 import {GameModel} from '../../models/game-model'
 import FireEffect from '../../status-effects/fire'
 import {beforeAttack} from '../../types/priorities'
-import {flipCoin} from '../../utils/coinFlips'
+import {flipCoin} from '../../utils/coin-flips'
 import {hermit} from '../defaults'
 import {Hermit} from '../types'
