+++ common/cards/hermits/humancleo-rare.ts
@@ -6,7 +6,7 @@ import {
 import {GameModel} from '../../models/game-model'
 import BetrayedEffect from '../../status-effects/betrayed'
 import {beforeAttack} from '../../types/priorities'
-import {flipCoin} from '../../utils/coinFlips'
+import {flipCoin} from '../../utils/coin-flips'
 import {hermit} from '../defaults'
 import {Hermit} from '../types'
