+++ common/cards/hermits/index.ts
@@ -20,7 +20,7 @@ import EthosLabUltraRare from './ethoslab-ultra-rare'
 import EvilJevinCommon from './eviljevin-common'
 import EvilJevinRare from './eviljevin-rare'
 import EvilXisumaCommon from './evilxisuma-common'
-import EvilXisumaRare from './evilxisuma_rare'
+import EvilXisumaRare from './evilxisuma-rare'
 import FalseSymmetryCommon from './falsesymmetry-common'
 import FalseSymmetryRare from './falsesymmetry-rare'
 import FarmerBeefCommon from './farmerbeef-common'
@@ -28,7 +28,7 @@ import FarmerBeefRare from './farmerbeef-rare'
 import FiveAMPearlCommon from './fiveampearl-common'
 import FiveAMPearlRare from './fiveampearl-rare'
 import FrenchKeralisCommon from './frenchkeralis-common'
-import FrenchralisRare from './frenchralis-rare'
+import FrenchralisRare from './frenchkeralis-rare'
 import GeminiTayCommon from './geminitay-common'
 import GeminiTayRare from './geminitay-rare'
 import GoatfatherCommon from './goatfather-common'
