+++ common/cards/hermits/xisumavoid-rare.ts
@@ -6,7 +6,7 @@ import {
 import {GameModel} from '../../models/game-model'
 import PoisonEffect from '../../status-effects/poison'
 import {beforeAttack} from '../../types/priorities'
-import {flipCoin} from '../../utils/coinFlips'
+import {flipCoin} from '../../utils/coin-flips'
 import {hermit} from '../defaults'
 import {Hermit} from '../types'
