+++ common/cards/single-use/bad-omen.ts
@@ -5,7 +5,7 @@ import {
 } from '../../components'
 import query from '../../components/query'
 import {GameModel} from '../../models/game-model'
-import BadOmenEffect from '../../status-effects/badomen'
+import BadOmenEffect from '../../status-effects/bad-omen'
 import {singleUse} from '../defaults'
 import {SingleUse} from '../types'
 
@@ -22,7 +22,7 @@ const BadOmen: SingleUse = {
 	sidebarDescriptions: [
 		{
 			type: 'statusEffect',
-			name: 'badomen',
+			name: 'bad-omen',
 		},
 	],
 	attachCondition: query.every(
