+++ common/cards/single-use/invisibility-potion.ts
@@ -8,7 +8,7 @@ import {
 	InvisibilityPotionHeadsEffect,
 	InvisibilityPotionTailsEffect,
 } from '../../status-effects/invisibility-potion'
-import {flipCoin} from '../../utils/coinFlips'
+import {flipCoin} from '../../utils/coin-flips'
 import {singleUse} from '../defaults'
 import {SingleUse} from '../types'
