+++ common/cards/single-use/looting.ts
@@ -1,7 +1,7 @@
 import {CardComponent, ObserverComponent, SlotComponent} from '../../components'
 import query from '../../components/query'
 import {GameModel} from '../../models/game-model'
-import {flipCoin} from '../../utils/coinFlips'
+import {flipCoin} from '../../utils/coin-flips'
 import {singleUse} from '../defaults'
 import {SingleUse} from '../types'
