+++ common/cards/single-use/spyglass.ts
@@ -1,7 +1,7 @@
 import {CardComponent, ObserverComponent} from '../../components'
 import query from '../../components/query'
 import {GameModel} from '../../models/game-model'
-import {flipCoin} from '../../utils/coinFlips'
+import {flipCoin} from '../../utils/coin-flips'
 import {getFormattedName} from '../../utils/game'
 import {singleUse} from '../defaults'
 import {SingleUse} from '../types'
