+++ common/cards/starter-decks.ts
@@ -20,10 +20,10 @@ import EthosLabCommon from './hermits/ethoslab-common'
 import EthosLabRare from './hermits/ethoslab-rare'
 import EthosLabUltraRare from './hermits/ethoslab-ultra-rare'
 import EvilJevinRare from './hermits/eviljevin-rare'
-import EvilXisumaRare from './hermits/evilxisuma_rare'
+import EvilXisumaRare from './hermits/evilxisuma-rare'
 import FiveAMPearlCommon from './hermits/fiveampearl-common'
 import FiveAMPearlRare from './hermits/fiveampearl-rare'
-import FrenchralisRare from './hermits/frenchralis-rare'
+import FrenchralisRare from './hermits/frenchkeralis-rare'
 import GeminiTayRare from './hermits/geminitay-rare'
 import GoatfatherRare from './hermits/goatfather-rare'
 import GrianCommon from './hermits/grian-common'
