+++ common/status-effects/bad-omen.ts
@@ -9,10 +9,10 @@ import {Counter, statusEffect} from './status-effect'
 
 const BadOmenEffect: Counter<CardComponent> = {
 	...statusEffect,
-	id: 'badomen',
-	icon: 'badomen',
+	id: 'bad-omen',
+	icon: 'bad-omen',
 	name: 'Bad Omen',
-	description: 'All coinflips are tails.',
+	description: 'All coin flips are tails.',
 	counter: 3,
 	counterType: 'turns',
