+++ common/status-effects/sheep-stare.ts
@@ -7,7 +7,7 @@ import {
 import query from '../components/query'
 import {GameModel} from '../models/game-model'
 import {afterAttack, beforeAttack, onTurnEnd} from '../types/priorities'
-import {flipCoin} from '../utils/coinFlips'
+import {flipCoin} from '../utils/coin-flips'
 import {StatusEffect, systemStatusEffect} from './status-effect'
 
 const SheepStareEffect: StatusEffect<PlayerComponent> = {
