+++ tests/replays/replays.test.ts
@@ -3,7 +3,7 @@ import Brush from 'common/cards/advent-of-tcg/single-use/brush'
 import Feather from 'common/cards/advent-of-tcg/single-use/feather'
 import {DiamondArmor} from 'common/cards/attach/armor'
 import EthosLabCommon from 'common/cards/hermits/ethoslab-common'
-import EvilXisumaRare from 'common/cards/hermits/evilxisuma_rare'
+import EvilXisumaRare from 'common/cards/hermits/evilxisuma-rare'
 import FalseSymmetryRare from 'common/cards/hermits/falsesymmetry-rare'
 import FarmerBeefRare from 'common/cards/hermits/farmerbeef-rare'
 import GeminiTayCommon from 'common/cards/hermits/geminitay-common'
