+++ tests/unit/game/advent-of-tcg/hermits/zookeeperscar-rare.test.ts
@@ -4,7 +4,7 @@ import ElderGuardian from 'common/cards/advent-of-tcg/attach/elder-guardian'
 import ZookeeperScarRare from 'common/cards/advent-of-tcg/hermits/zookeeperscar-rare'
 import Wolf from 'common/cards/attach/wolf'
 import EthosLabCommon from 'common/cards/hermits/ethoslab-common'
-import EvilXisumaRare from 'common/cards/hermits/evilxisuma_rare'
+import EvilXisumaRare from 'common/cards/hermits/evilxisuma-rare'
 import RendogRare from 'common/cards/hermits/rendog-rare'
 import BalancedItem from 'common/cards/items/balanced-common'
 import Emerald from 'common/cards/single-use/emerald'
